feat(resultify): add resultify function to convert boolean values to Result type
The `resultify` function was added to convert boolean values to the `Result` type. This function receives a function as a parameter and returns a `Promise` that resolves to a `Result` type. If the function returns a truthy value, the `Result` will be `ok` with the value returned by the function. If the function returns a falsy value, the `Result` will be `error` with an `Error` object containing the message "The boolean value returns false". If the function throws an error, the `Result` will be `error` with the thrown error.